Split toPencil into two funcutions
The preivous commit stripped toPencil out of most computations because
the results were not used.  The remaining code computes (1) a
parameter-value map and (2) metadata of the output tensors.  Split this
function into two individual functions, computeParamValueMap for (1) and
inferOutputTensorInfo for (2), so that they can be called independently.
Update CudaTcExecutor to remove getPvmAndOutputTensorInfo that became a
forwarding call to computeParamValueMap.