halide2isl: make sure extractAccess is called once on each IRNode
extractAccess introduces reference identifiers with sequential numbers
in the domain of the access relation it constructs.  The reference
idetifiers are stored in a map halide::IRNode->isl::id.  Assuming that
each IRNode is visited once by extractAccess, the sequential numbers are
obtained by querying the size of the map.  However, some IRNodes may be
visited more than once, for example, with let statement substitution.
Make FindAccesses visitor class, which class extractAccess on IRNodes,
inherit IRGraphVisitor instead of IRVisitor to ensure each IRNode is
visited once.

This fixes the problem that appeared in memory promotion when the same
reference identifier was attached to different tensor names.